Split draw::primitive::{path, polygon} into separate modules
