+++ nannou/src/draw/primitive/path/mod.rs
@@ -1,6 +1,12 @@
+pub mod render;
+
+pub(crate) use self::render::{
+    render_path_events, render_path_points_colored, render_path_points_textured,
+    render_path_source, PathEventSource, PathEventSourceIter,
+};
 use crate::color::conv::IntoLinSrgba;
 use crate::color::LinSrgba;
-use crate::draw::mesh::vertex::{Color, TexCoords};
+use crate::draw::mesh::vertex::TexCoords;
 use crate::draw::primitive::Primitive;
 use crate::draw::properties::spatial::{orientation, position};
 use crate::draw::properties::{
@@ -8,7 +14,6 @@ use crate::draw::properties::{
 };
 use crate::draw::{self, Drawing, DrawingContext};
 use crate::geom::Point2;
-use crate::glam::Mat4;
 use crate::wgpu;
 use lyon::path::PathEvent;
 use lyon::tessellation::{FillOptions, FillTessellator, StrokeOptions, StrokeTessellator};
@@ -19,34 +24,6 @@ pub trait TessellationOptions {
     type Tessellator;
     /// Convert the typed options instance to a dynamic one.
     fn into_options(self) -> Options;
-}
-
-#[derive(Clone, Debug)]
-pub(crate) enum PathEventSource {
-    /// Fetch events from `path_events_buffer`.
-    Buffered(std::ops::Range<usize>),
-    /// Generate events from the `path_points_colored_buffer`.
-    ColoredPoints {
-        range: std::ops::Range<usize>,
-        close: bool,
-    },
-    /// Generate events from the `path_points_textured_buffer`.
-    TexturedPoints {
-        range: std::ops::Range<usize>,
-        close: bool,
-    },
-}
-
-pub(crate) enum PathEventSourceIter<'a> {
-    Events(&'a mut dyn Iterator<Item = lyon::path::PathEvent>),
-    ColoredPoints {
-        points: &'a mut dyn Iterator<Item = (Point2, Color)>,
-        close: bool,
-    },
-    TexturedPoints {
-        points: &'a mut dyn Iterator<Item = (Point2, TexCoords)>,
-        close: bool,
-    },
 }
 
 /// The beginning of the path building process, prior to choosing the tessellation mode (fill or
@@ -349,309 +326,6 @@ where
             Some(texture_view),
         )
     }
-}
-
-pub(crate) fn render_path_events<I>(
-    events: I,
-    color: Option<LinSrgba>,
-    transform: Mat4,
-    options: Options,
-    theme: &draw::Theme,
-    theme_prim: &draw::theme::Primitive,
-    fill_tessellator: &mut lyon::tessellation::FillTessellator,
-    stroke_tessellator: &mut lyon::tessellation::StrokeTessellator,
-    mesh: &mut draw::Mesh,
-) where
-    I: IntoIterator<Item = lyon::path::PathEvent>,
-{
-    let res = match options {
-        Options::Fill(options) => {
-            let color = color.unwrap_or_else(|| theme.fill_lin_srgba(theme_prim));
-            let mut mesh_builder = draw::mesh::MeshBuilder::single_color(mesh, transform, color);
-            fill_tessellator.tessellate(events, &options, &mut mesh_builder)
-        }
-        Options::Stroke(options) => {
-            let color = color.unwrap_or_else(|| theme.stroke_lin_srgba(theme_prim));
-            let mut mesh_builder = draw::mesh::MeshBuilder::single_color(mesh, transform, color);
-            stroke_tessellator.tessellate(events, &options, &mut mesh_builder)
-        }
-    };
-    if let Err(err) = res {
-        eprintln!("failed to tessellate path: {:?}", err);
-    }
-}
-
-pub(crate) fn render_path_points_colored<I>(
-    points_colored: I,
-    close: bool,
-    transform: Mat4,
-    options: Options,
-    fill_tessellator: &mut lyon::tessellation::FillTessellator,
-    stroke_tessellator: &mut lyon::tessellation::StrokeTessellator,
-    mesh: &mut draw::Mesh,
-) where
-    I: IntoIterator<Item = (Point2, Color)>,
-{
-    let path = match points_colored_to_lyon_path(points_colored, close) {
-        None => return,
-        Some(p) => p,
-    };
-
-    // Extend the mesh with the built path.
-    let mut mesh_builder = draw::mesh::MeshBuilder::color_per_point(mesh, transform);
-    let res = match options {
-        Options::Fill(options) => fill_tessellator.tessellate_with_ids(
-            path.id_iter(),
-            &path,
-            Some(&path),
-            &options,
-            &mut mesh_builder,
-        ),
-        Options::Stroke(options) => stroke_tessellator.tessellate_with_ids(
-            path.id_iter(),
-            &path,
-            Some(&path),
-            &options,
-            &mut mesh_builder,
-        ),
-    };
-    if let Err(err) = res {
-        eprintln!("failed to tessellate path: {:?}", err);
-    }
-}
-
-pub(crate) fn render_path_points_textured<I>(
-    points_textured: I,
-    close: bool,
-    transform: Mat4,
-    options: Options,
-    fill_tessellator: &mut lyon::tessellation::FillTessellator,
-    stroke_tessellator: &mut lyon::tessellation::StrokeTessellator,
-    mesh: &mut draw::Mesh,
-) where
-    I: IntoIterator<Item = (Point2, TexCoords)>,
-{
-    let path = match points_textured_to_lyon_path(points_textured, close) {
-        None => return,
-        Some(p) => p,
-    };
-
-    // Extend the mesh with the built path.
-    let mut mesh_builder = draw::mesh::MeshBuilder::tex_coords_per_point(mesh, transform);
-    let res = match options {
-        Options::Fill(options) => fill_tessellator.tessellate_with_ids(
-            path.id_iter(),
-            &path,
-            Some(&path),
-            &options,
-            &mut mesh_builder,
-        ),
-        Options::Stroke(options) => stroke_tessellator.tessellate_with_ids(
-            path.id_iter(),
-            &path,
-            Some(&path),
-            &options,
-            &mut mesh_builder,
-        ),
-    };
-    if let Err(err) = res {
-        eprintln!("failed to tessellate path: {:?}", err);
-    }
-}
-
-pub(crate) fn render_path_source(
-    // TODO:
-    path_src: PathEventSourceIter,
-    color: Option<LinSrgba>,
-    transform: Mat4,
-    options: Options,
-    theme: &draw::Theme,
-    theme_prim: &draw::theme::Primitive,
-    fill_tessellator: &mut lyon::tessellation::FillTessellator,
-    stroke_tessellator: &mut lyon::tessellation::StrokeTessellator,
-    mesh: &mut draw::Mesh,
-) {
-    match path_src {
-        PathEventSourceIter::Events(events) => render_path_events(
-            events,
-            color,
-            transform,
-            options,
-            theme,
-            theme_prim,
-            fill_tessellator,
-            stroke_tessellator,
-            mesh,
-        ),
-        PathEventSourceIter::ColoredPoints { points, close } => render_path_points_colored(
-            points,
-            close,
-            transform,
-            options,
-            fill_tessellator,
-            stroke_tessellator,
-            mesh,
-        ),
-        PathEventSourceIter::TexturedPoints { points, close } => render_path_points_textured(
-            points,
-            close,
-            transform,
-            options,
-            fill_tessellator,
-            stroke_tessellator,
-            mesh,
-        ),
-    }
-}
-
-impl draw::renderer::RenderPrimitive for Path {
-    fn render_primitive(
-        self,
-        mut ctxt: draw::renderer::RenderContext,
-        mesh: &mut draw::Mesh,
-    ) -> draw::renderer::PrimitiveRender {
-        let Path {
-            color,
-            position,
-            orientation,
-            path_event_src,
-            options,
-            vertex_mode,
-            texture_view,
-        } = self;
-
-        // Determine the transform to apply to all points.
-        let global_transform = *ctxt.transform;
-        let local_transform = position.transform() * orientation.transform();
-        let transform = global_transform * local_transform;
-
-        // A function for rendering the path.
-        let render =
-            |src: PathEventSourceIter,
-             theme: &draw::Theme,
-             fill_tessellator: &mut lyon::tessellation::FillTessellator,
-             stroke_tessellator: &mut lyon::tessellation::StrokeTessellator| {
-                render_path_source(
-                    src,
-                    color,
-                    transform,
-                    options,
-                    theme,
-                    &draw::theme::Primitive::Path,
-                    fill_tessellator,
-                    stroke_tessellator,
-                    mesh,
-                )
-            };
-
-        match path_event_src {
-            PathEventSource::Buffered(range) => {
-                let mut events = ctxt.path_event_buffer[range].iter().cloned();
-                let src = PathEventSourceIter::Events(&mut events);
-                render(
-                    src,
-                    &ctxt.theme,
-                    &mut ctxt.fill_tessellator,
-                    &mut ctxt.stroke_tessellator,
-                );
-            }
-            PathEventSource::ColoredPoints { range, close } => {
-                let mut points_colored = ctxt.path_points_colored_buffer[range].iter().cloned();
-                let src = PathEventSourceIter::ColoredPoints {
-                    points: &mut points_colored,
-                    close,
-                };
-                render(
-                    src,
-                    &ctxt.theme,
-                    &mut ctxt.fill_tessellator,
-                    &mut ctxt.stroke_tessellator,
-                );
-            }
-            PathEventSource::TexturedPoints { range, close } => {
-                let mut points_textured = ctxt.path_points_textured_buffer[range].iter().cloned();
-                let src = PathEventSourceIter::TexturedPoints {
-                    points: &mut points_textured,
-                    close,
-                };
-                render(
-                    src,
-                    &ctxt.theme,
-                    &mut ctxt.fill_tessellator,
-                    &mut ctxt.stroke_tessellator,
-                );
-            }
-        }
-
-        draw::renderer::PrimitiveRender {
-            texture_view,
-            vertex_mode,
-        }
-    }
-}
-
-/// Create a lyon path for the given iterator of colored points.
-pub fn points_colored_to_lyon_path<I>(points_colored: I, close: bool) -> Option<lyon::path::Path>
-where
-    I: IntoIterator<Item = (Point2, Color)>,
-{
-    // Build a path with a color attribute for each channel.
-    let channels = draw::mesh::vertex::COLOR_CHANNEL_COUNT;
-    let mut path_builder = lyon::path::Path::builder_with_attributes(channels);
-
-    // Begin the path.
-    let mut iter = points_colored.into_iter();
-    let (first_point, first_color) = iter.next()?;
-    let p = first_point.to_array().into();
-    let (r, g, b, a) = first_color.into();
-    path_builder.move_to(p, &[r, g, b, a]);
-
-    // Add the lines, keeping track of the last
-    for (point, color) in iter {
-        let p = point.to_array().into();
-        let (r, g, b, a) = color.into();
-        path_builder.line_to(p, &[r, g, b, a]);
-    }
-
-    // Close if necessary.
-    if close {
-        path_builder.close();
-    }
-
-    // Build it!
-    Some(path_builder.build())
-}
-
-/// Create a lyon path for the given iterator of textured points.
-pub fn points_textured_to_lyon_path<I>(points_textured: I, close: bool) -> Option<lyon::path::Path>
-where
-    I: IntoIterator<Item = (Point2, TexCoords)>,
-{
-    // Build a path with a texture coords attribute for each channel.
-    let channels = 2;
-    let mut path_builder = lyon::path::Path::builder_with_attributes(channels);
-
-    // Begin the path.
-    let mut iter = points_textured.into_iter();
-    let (first_point, first_tex_coords) = iter.next()?;
-    let p = first_point.to_array().into();
-    let (tc_x, tc_y) = first_tex_coords.into();
-    path_builder.move_to(p, &[tc_x, tc_y]);
-
-    // Add the lines, keeping track of the last
-    for (point, tex_coords) in iter {
-        let p = point.to_array().into();
-        let (tc_x, tc_y) = tex_coords.into();
-        path_builder.line_to(p, &[tc_x, tc_y]);
-    }
-
-    // Close if necessary.
-    if close {
-        path_builder.close();
-    }
-
-    // Build it!
-    Some(path_builder.build())
 }
 
 impl Path {
